+++ actix-tls/src/connect/error.rs
@@ -1,15 +1,16 @@
-use std::io;
+use std::{error::Error, io};
 
 use derive_more::Display;
 
+/// Errors that can result from using a connector service.
 #[derive(Debug, Display)]
 pub enum ConnectError {
     /// Failed to resolve the hostname
-    #[display(fmt = "Failed resolving hostname: {}", _0)]
+    #[display(fmt = "Failed resolving hostname")]
     Resolver(Box<dyn std::error::Error>),
 
-    /// No dns records
-    #[display(fmt = "No dns records found for the input")]
+    /// No DNS records
+    #[display(fmt = "No DNS records found for the input")]
     NoRecords,
 
     /// Invalid input
@@ -22,4 +23,14 @@ pub enum ConnectError {
     /// Connection IO error
     #[display(fmt = "{}", _0)]
     Io(io::Error),
+}
+
+impl Error for ConnectError {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
+        match self {
+            Self::Resolver(err) => Some(&**err),
+            Self::Io(err) => Some(err),
+            Self::NoRecords | Self::InvalidInput | Self::Unresolved => None,
+        }
+    }
 }